Muda referência recebida pela URL, troca .../username/... para .../user.id/...

+++ website/views.py
@@ -4,11 +4,11 @@ from django.contrib.auth.decorators import login_required, permission_required
 from django.contrib.auth import authenticate
 from django.contrib.auth import login as login_user
 from django.contrib.auth import logout as logout_user
-
 from .models import *
 from .forms import *
 
 def home(request):
+	print(str(request.user.user_permissions))
 	return render(request, 'home.html')
 
 def login(request):
@@ -152,7 +152,7 @@ def competition_details(request, competition_id):
 		'title': competition,
 		'competition': competition,
 		'tournament': tournament,
-		'trials': competition.matches.filter(intercampi=False),
+		'trials': competition.matchs.filter(intercampi=False),
 		'intercampi': Match.objects.filter(competition=competition, intercampi=True).first(),
 		'breadcrumb': [
 			{'name': 'Início', 'link': '/'},
@@ -221,6 +221,8 @@ def match_details(request, match_id):
 @login_required()
 def edit_match(request, match_id):
 	match = get_object_or_404(Match, id=match_id)
+	if request.user.username!=match.responsible.username:
+		return redirect('/')
 	if request.method == 'POST':
 		form = MatchForm(request.POST, instance=match)
 		if form.is_valid():
@@ -228,6 +230,7 @@ def edit_match(request, match_id):
 			return redirect('/jogos/' + match_id)
 	else:
 		form = MatchForm(instance=match)
+
 
 	competition = get_object_or_404(Competition, id=match.competition.id)
 	context = {
@@ -252,6 +255,8 @@ def edit_match(request, match_id):
 @login_required()
 def remove_match(request, match_id):
 	match = get_object_or_404(Match, id=match_id)
+	if request.user.username!=match.responsible.username:
+		return redirect('/')
 	competition_id = match.competition.id
 	match.delete()
 	return redirect('/competicoes/' + str(competition_id))
@@ -297,7 +302,7 @@ def attend_to_match(request, match_id):
 @login_required()
 def leave_match(request, match_id):
 	match = get_object_or_404(Match, id=match_id)
-	match.participants.remove(request.user)
+	match.teams.remove(request.user)
 	score = MatchScore.objects.get(match=match, participant=request.user)
 	score.delete()
 	return redirect('/jogos/' + str(match_id))
@@ -337,14 +342,13 @@ def list_results(request):
 		not_matchs = Match.match_not_ready(request) + Match.matchs_ready_to_publish_result(request)
 	for match in matchs:
 			matchScore = MatchScore.objects.all().filter(match=match)
-			if match.participants.count()!=matchScore.count():
+			if match.teams.count()!=matchScore.count():
 				match.first_place=None
 				match.save()
-			elif match.participants.count()==matchScore.count() and match.first_place:
+			elif match.teams.count()==matchScore.count() and match.first_place:
 				boolean = True
 	context = {
 		'title': 'Resultados',
-		'USER': request.user.username,
 		'matchs': matchs,
 		'is_complete': boolean,
 		'not_matchs': not_matchs,
@@ -356,32 +360,34 @@ def list_results(request):
 	return render(request, 'list-results.html', context)
 
 @login_required
-def list_incomplete_or_not_plubished_results(request, user_name):
-	if request.user.username != user_name:
+def list_incomplete_or_not_plubished_results(request, user_id):
+	user = User.objects.all().filter(id=user_id).first()
+	if request.user != user:
 		return redirect('/resultados')
 	MATCHS = Match.matchs_ready_to_publish_result(request)
 	MATCH  = Match.match_not_ready(request)
+	MATC   = Match.match_already_published(request)
 	context = {
-		'title': 'Partidas - '+ user_name,
-		'USER': request.user.username,
-		'USER_MORE': user_name,
+		'title': 'Partidas - '+ request.user.get_full_name(),
 		'matchs': MATCHS,
 		'matchs_not_ready': MATCH,
+		'match_already_published': MATC,
 		'breadcrumb': [
 			{'name': 'Início', 'link': '/'},
-			{'name': 'Torneios', 'link': '/torneios'},
-			{'name': 'Competições', 'link': '/competicoes'},
-			{'name': 'Jogo', 'link': '/jogos'},
 			{'name': 'Resultados', 'link': '/resultados'},
+			{'name': user.first_name},
 		]
 	}
 
 	return render(request, 'list-incomplete-results.html', context)
 
-def match_score(request, user_name, match_id):
+def match_score(request, user_id, match_id):
 	match = get_object_or_404(Match, id=match_id)
+	user = User.objects.all().filter(id=user_id).first()
+	if user.username!=match.responsible.username:
+		return redirect("/resultados")
 	matchScore = MatchScore.objects.all().filter(match=match)
-	if not matchScore.count() == match.participants.count() or not match.first_place:
+	if not matchScore.count() == match.teams.count() or not match.first_place:
 		if request.user.is_anonymous():
 			return redirect('/resultados')
 
@@ -390,8 +396,7 @@ def match_score(request, user_name, match_id):
 		'title': 'Resultado ' + match.type() + ' - ' + str(match.competition),
 		'matchScores': matchScore,
 		'match':match,
-		'USER': request.user.username,
-		'USER_MOR': match.responsible.username,
+		'USER_MOR': match.responsible,
 		'breadcrumb': [
 			{'name': 'Início', 'link': '/'},
 			{'name': 'Torneios', 'link': '/torneios'},
@@ -400,67 +405,58 @@ def match_score(request, user_name, match_id):
 			{'name': match.competition, 'link': '/competicoes/' + str(match.competition.id)},
 			{'name': match.type() },
 			{'name': 'Resultados', 'link': '/resultados'},
-			{'name': match.responsible.username, 'link': '/resultados/' + match.responsible.username},
+			{'name': match.responsible.first_name, 'link': '/resultados/' + str(match.responsible.id)},
 		]
 	}
 	return render(request, 'match-score-details.html', context)
 
 @login_required
-def publish_result(request, user_name, match_id):
-	if request.user.username != user_name:
+def publish_result(request, user_id, match_id):
+	user = User.objects.all().filter(id=user_id).first()
+	if request.user != user:
 		return redirect('/resultados')
 	match = get_object_or_404(Match, id=match_id)
-						# Pega todos os MatchScore da partida, ordena de maneira 
-						# decrescente e pega o primeiro (Maior pontuação)
-	match.first_place = MatchScore.objects.all().filter(match=match).order_by('-score').first().participant
+	match.first_place = MatchScore.objects.all().filter(match=match).order_by('-score').first().team
 	match.save()
-	return redirect('/resultados/'+ user_name + '/' +match_id)
+	return redirect('/resultados/'+str(user_id))
 
 @login_required
-def list_incomplete_scores(request, user_name, match_id):
-	if request.user.username != user_name:
+def list_incomplete_scores(request, user_id, match_id):
+	user = User.objects.all().filter(id=user_id).first()
+	if request.user != user:
 		return redirect('/resultados')
 	match = get_object_or_404(Match, id=match_id)
 	scores = MatchScore.objects.all().filter(match = match)
-	participants_complete_score = []
-	participants_incomplete_score = []
+	teams_incomplete_score = []
 	
-	for score in scores:                        #Percorre MatchScore já cadastrados
-		for participant in Participant.objects.all():  #Percorre Participantes da match
-			if score.participant == participant:#Se Participante == MatchScore.participante 
-				participants_complete_score.append(participant)		#Adiciona a matchScore na lista de MS completa
-
-	for participant in match.participants.all():
-		matchScores = scores.filter(participant=participant)
+	for team in match.teams.all():
+		matchScores = scores.filter(team=team)
 		if not matchScores:
-			participants_incomplete_score.append(participant)
+			teams_incomplete_score.append(team)
 	
 	context = {
-		'title': 'Pontuação Individual - '+ str(match),
-		'USER': request.user.username,
-		'USER_MORE': user_name,
+		'title': 'Pontuação Individual',
 		'match': match,
 		'matchScores': scores.order_by('-score'),
-		'participants_without_matchScore': participants_incomplete_score,
+		'teams_without_matchScore': teams_incomplete_score,
 
 		'breadcrumb': [
 			{'name': 'Início', 'link': '/'},
-			{'name': 'Torneios', 'link': '/torneios'},
-			{'name': 'Competições', 'link': '/competicoes'},
-			{'name': 'Jogo', 'link': '/jogos'},
 			{'name': 'Resultados', 'link': '/resultados'},
-			{'name': user_name, 'link': '/resultados/' + user_name},
+			{'name': user.first_name, 'link': '/resultados/' + str(user_id)},
+			{'name': 'Pontuação - ' + str(match)},
 		]
 	}
 
 	return render(request, 'matchscore-add.html', context)
 
 @login_required
-def add_matchScore(request, user_name, match_id, participant_id):
-	if request.user.username != user_name:
+def add_matchScore(request, user_id, match_id, team_id):
+	user = User.objects.all().filter(id=user_id).first()
+	if request.user != user:
 		return redirect('/resultados')
 	match = get_object_or_404(Match, id=match_id)
-	participant = get_object_or_404(Participant, id=participant_id)
+	team = get_object_or_404(Team, id=team_id)
 	competition = match.competition
 	tournament = competition.tournament
 
@@ -468,30 +464,73 @@ def add_matchScore(request, user_name, match_id, participant_id):
 		form = MatchScoreForm(request.POST)
 		if form.is_valid():
 			matchScore = form.save(commit=False)
+			team.score = matchScore.score
+			team.save()
 			match.first_place=None
 			match.save()
 			matchScore.match=match
-			matchScore.participant=participant
+			matchScore.team=team
 			matchScore.save()
-			return redirect("/pontuacao/" + user_name + '/' + str(match_id) + '/')
+			return redirect("/pontuacao/" + str(user_id) + '/' + str(match_id) + '/')
 	else:
 		form = MatchScoreForm()
 
 	context = {
-		'title': "Pontuação" + str(match) + ' - ' + participant.name,
-		'action': '/pontuacao/' + user_name + '/' + str(match_id) + '/' + str(participant_id) + '/',
-		'cancel': '/pontuacao/' + user_name + '/' + str(match_id),
+		'title': "Pontuação - " + team.name,
+		'action': '/pontuacao/' + str(user_id) + '/' + str(match_id) + '/' + str(team_id) + '/',
+		'cancel': '/pontuacao/' + str(user_id) + '/' + str(match_id),
 		'form': form,
 		'competition' : competition,
 		'breadcrumb': [
 			{'name': 'Início', 'link': '/'},
-			{'name': 'Competições', 'link': '/torneios/' + str(tournament.id)},
-			{'name': competition, 'link': '/competitions/' + str(competition.id)},
-			{'name': match, 'link': '/jogos/' + str(match_id)},
 			{'name': 'Resultados', 'link': '/resultados'},
-			{'name': 'Resultados ' + str(match), 'link': '/resultados/' + user_name + '/' + str(match_id)},
+			{'name': user.first_name, 'link': '/resultados/' + str(user_id)},
+			{'name': 'Pontuação - ' + str(match), 'link': '/pontuacao/' + str(user_id) + '/' + str(match_id)},
 			{'name': 'Nova'},
 		]
 	}
 
 	return render(request, 'form.html', context)
+
+@login_required
+def edit_matchScore(request, user_id, matchScore_id):
+	user = User.objects.all().filter(id=user_id).first()
+	if request.user != user:
+		return redirect("/resultados")
+	matchScore = get_object_or_404(MatchScore, id=matchScore_id)
+
+	if request.method == 'POST':
+		form = MatchScoreForm(request.POST, instance=matchScore)
+		if form.is_valid():
+			matchScore.match.first_place=None
+			matchScore.match.save()
+			form.save()
+			return redirect('/pontuacao/' + str(user_id) + '/' + str(matchScore.match.id))
+	else:
+		form = MatchScoreForm(instance=matchScore)
+
+	context = {
+		'title': "Pontuação - " + str(matchScore.match) + ' - ' + matchScore.team.name,
+		'action': '/pontuacao/editar/' + str(user_id) + '/' + str(matchScore_id) + '/',
+		'cancel': '/pontuacao/' + str(user_id) + '/' + str(matchScore.match.id),
+		'form': form,
+		'breadcrumb': [
+			{'name': 'Início', 'link': '/'},
+			{'name': 'Resultados', 'link': '/resultados'},
+			{'name': user.first_name, 'link': '/resultados/' + str(user_id)},
+			{'name': 'Pontuação - ' + str(match), 'link': '/pontuacao/' + str(user_id) + '/' + str(match_id)},
+			{'name': 'Editar'},
+		]
+	}
+
+	return render(request, 'form.html', context)
+
+@login_required
+def remove_matchScore(request, user_id,matchScore_id):
+	user = User.objects.all().filter(id=user_id).first()
+	if request.user!=user:
+		return redirect('/')
+	matchScore = get_object_or_404(MatchScore, id=matchScore_id)
+	match_id = matchScore.match.id
+	matchScore.delete()
+	return redirect('/pontuacao/' + str(user_id) + '/' + str(match_id))
